+++ src/utils/descriptors/descriptors.cpp
@@ -164,6 +164,10 @@ namespace Engine {
 
     void DescriptorWriter::overwrite(VkDescriptorSet &set) {
         for (auto &write : writes) write.dstSet = set;
-        vkUpdateDescriptorSets(pool.device.device(), writes.size(), writes.data(), 0, nullptr);
+        vkUpdateDescriptorSets(pool.device.device(),
+                               static_cast<unsigned int>(writes.size()),
+                               writes.data(),
+                               0,
+                               nullptr);
     }
 }
